Move get_template_parameters_s3 to ef_conf_utils
